% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dim_channel}
\alias{dim_channel}
\title{Demo data: Dim Channel}
\format{
A data frame with 616 rows and 5 columns: \describe{
  \item{channel_key}{Primary key (unique identifyer of channels).}
  \item{channel_grouping}{The main type of channel, e.g. Email, Social.}
  \item{source}{The source of referrals.}
  \item{medium}{The type of referrals.}
  \item{campaign}{The name of the campaign.} }
}
\source{
Anonymized data from google analytics.
}
\usage{
dim_channel
}
\description{
A look-up dimension related to Fact Web Sales through the channel_key column.
}
\keyword{datasets}
