% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dim_hour}
\alias{dim_hour}
\title{Demo data: Dim Hour}
\format{
A data frame with 24 rows and 2 columns:
\describe{
  \item{hour_key}{Primary key (unique identifier of hour).}
  \item{hour}{Hour as a name (character type).}
  }
}
\source{
Anonymized data from google analytics.
}
\usage{
dim_hour
}
\description{
A look-up dimension related to Fact Visitors through the hour_key column.
}
\keyword{datasets}
