% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dim_market}
\alias{dim_market}
\title{Demo data: Dim Market}
\format{
A data frame with 2 rows and 3 columns:
\describe{
  \item{market_key}{Primary key (unique identifyer of markets).}
  \item{view_name}{The name of the Google Analytics View.}
  \item{country}{The country associated with the Google Analytics View.}
  }
}
\source{
Anonymized data from google analytics.
}
\usage{
dim_market
}
\description{
A look-up dimension related to Fact Web Sales through the market_key column.
}
\keyword{datasets}
