% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fact_visitors}
\alias{fact_visitors}
\title{Demo data: Fact Visitors}
\format{
A data frame with 10,033 rows and 5 columns:
\describe{
  \item{visitor_id}{Unique identifier of the visitor.}
  \item{transaction_id}{Unique identifier of the transactions.}
  \item{revenue}{The value of the transaction in USD.}
  \item{timestamp}{The time of visit in minutes since 1970-01-01.}
  \item{hour_key}{Foreign key referring to dim_hour.}
  }
}
\source{
Anonymized data from google analytics.
}
\usage{
fact_visitors
}
\description{
A fact table showing individual visitors and their transactions on an
e-commerce website.
}
\keyword{datasets}
