% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fact_websales}
\alias{fact_websales}
\title{Demo data: Fact Web Sales}
\format{
A data frame with 1,500 rows and 9 columns:
\describe{
  \item{date}{The date of the fact.}
  \item{market_key}{Foreign key referring to dim_market.}
  \item{channel_key}{Foreign key referring to dim_channel}
  \item{ad_impressions}{The number of impressions of the associated ad.}
  \item{ad_clicks}{The number of times a user has cliked on the associated
  ad.}
  \item{ad_cost}{The total cost of the ad in USD.}
  \item{sessions}{The number of visits from the associated channel.}
  \item{transactions}{The number of purchases mediated by the associated
  channel.}
  \item{revenue}{The total value of purchases mediated by the associated
  channel.}
  }
}
\source{
Anonymized data from google analytics.
}
\usage{
fact_websales
}
\description{
A fact table showing number of visits, transactions, revenue and advertising
metrics by date, channel and market.
}
\keyword{datasets}
