% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{pbi_dataset_refresh}
\alias{pbi_dataset_refresh}
\title{Refresh dataset}
\usage{
pbi_dataset_refresh(
  group_id,
  dataset_id,
  notify = c("MailOnFailure", "MailOnCompletion", "NoNotification")
)
}
\arguments{
\item{group_id}{The ID of the workspace.}

\item{dataset_id}{The ID of the dataset.}

\item{notify}{Mail notification options. Must be one of 'MailOnFailure',
'MailOnCompletion', 'NoNotification'. Defaults to 'MailOnFailure'.}
}
\value{
The specified dataset will be refreshed.
}
\description{
Triggers a refresh for the specified dataset from the specified workspace.
}
\examples{

\dontrun{

pbi_dataset_refresh(group_id, dataset_id)
}
}
