% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{pbi_dataset_refresh_hist}
\alias{pbi_dataset_refresh_hist}
\title{Refresh history of a dataset}
\usage{
pbi_dataset_refresh_hist(group_id, dataset_id)
}
\arguments{
\item{group_id}{The workspace ID}

\item{dataset_id}{The dataset ID}
}
\value{
A data frame
}
\description{
Returns the refresh history for the specified dataset from the specified workspace.
}
\examples{

\dontrun{

pbi_dataset_refresh_hist(group_id, dataset_id)
}
}
