% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{pbi_delete_dataset}
\alias{pbi_delete_dataset}
\title{Title}
\usage{
pbi_delete_dataset(group_id, dataset_id)
}
\arguments{
\item{group_id}{Grp id}

\item{dataset_id}{test}
}
\value{
Deleted rows
}
\description{
My description
}
\examples{

\dontrun{

pbi_delete_dataset(group_id, dataset_id)
}
}
