% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{pbi_delete_rows}
\alias{pbi_delete_rows}
\title{Title}
\usage{
pbi_delete_rows(group_id, dataset_id, table_name)
}
\arguments{
\item{group_id}{The grp id}

\item{dataset_id}{The dataset id}

\item{table_name}{The table name}
}
\value{
Rows will be deleted
}
\description{
This is my description.
}
\examples{

\dontrun{

pbi_delete_rows(group_id, dataset_id, table_name)
}
}
