% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{pbi_list_datasets}
\alias{pbi_list_datasets}
\title{Get a list of datasets in a Power BI workspace}
\usage{
pbi_list_datasets(group_id)
}
\arguments{
\item{group_id}{The Power BI workspace ID.}
}
\value{
A data.table / data frame with dataset information.
}
\description{
Returns the IDs and meta data of all available datasets in the specified
Power BI workspace.
}
\details{
Required scope: Dataset.ReadWrite.All or Dataset.Read.All
}
\examples{

\dontrun{

pbi_list_datasets(my_group_id)
}
}
