% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{pbi_list_groups}
\alias{pbi_list_groups}
\title{List Power BI groups (workspaces)}
\usage{
pbi_list_groups()
}
\value{
A data.table with Power BI group ids and meta data.
}
\description{
List Power BI groups (workspaces)
}
