% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{pbi_push_dataset}
\alias{pbi_push_dataset}
\title{Push Dataset}
\usage{
pbi_push_dataset(schema, group_id, retention = c("none", "basicFIFO"))
}
\arguments{
\item{schema}{A push-dataset schema created by pbi_schema_create().}

\item{group_id}{The ID of the destination Power BI workspace.}

\item{retention}{The retention policy of the dataset. Default is "none".}
}
\value{
A dataset with tables and optionally defined relationships will be
  created in the specified Power BI workspace.
}
\description{
Pushes an empty dataset to the specified Power BI workspace. To add rows to
the dataset, use pbi_push_rows().
}
\details{
Required scope: Dataset.ReadWrite.All
}
\examples{

\dontrun{

pbi_push_dataset(my_schema, my_group_id)
}
}
