% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/push.R
\name{pbi_push_rows}
\alias{pbi_push_rows}
\title{Title}
\usage{
pbi_push_rows(dt, group_id, dataset_id, table_name)
}
\arguments{
\item{dt}{A data.table}

\item{group_id}{The grp id, workspace}

\item{dataset_id}{The dataset}

\item{table_name}{The name of the table}
}
\value{
A dataset with tables and optionally defined relationships will be
  created in the specified Power BI workspace
}
\description{
My description
}
\examples{

\dontrun{

pbi_push_rows(my_data, my_group_id)
}
}
