% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{pbi_schema_add_relations}
\alias{pbi_schema_add_relations}
\title{Define Power BI table relations}
\usage{
pbi_schema_add_relations(schema, rel_list)
}
\arguments{
\item{schema}{A schema object from pbi_schema_create()}

\item{rel_list}{A list of relations.}
}
\value{
An updated schema including relations definitions.
}
\description{
Define Power BI table relations
}
\examples{

# Dummy data: Two data frames with shared ID columns.
df1 <- swiss
df1$id <- rownames(df1)

df2 <- swiss
df2$id <- rownames(df2)
df2$var <- rnorm(nrow(df2), 10, 4)
df2 <- df2[, c("id", "var" )]

# Create schema
table_list <- list(df1, df2)
table_names  <- c("DF1", "DF2")
dataset_name <- "MyDataSet"

schema <- pbi_schema_create(
  dt_list = table_list,
  dataset_name = dataset_name,
  table_name_list = table_names
  )

# Add relations definitions
relations <- data.frame(
  name = "MyUniqueRelationsID",
  fromTable = "swiss",
  fromColumn = "id",
  toTable = "swiss_dummy",
  toColumn = "id",
  crossFilteringBehavior = "bothDirections"
  )

# Add relations to schema
relations_list <- list(relations)

schema <- pbi_schema_add_relations(
  schema = schema,
  rel_list = relations_list
  )
}
