% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{pbi_schema_column_prop}
\alias{pbi_schema_column_prop}
\title{Set Power BI column properties}
\usage{
pbi_schema_column_prop(
  dt,
  table_name = "table1",
  date_format = "yyyy-mm-dd",
  integer_format = "#,###0",
  double_format = "#,###.00"
)
}
\arguments{
\item{dt}{A data frame.}

\item{table_name}{The name given to the data frame. Default is 'table1'.}

\item{date_format}{How the datetime columns should be formatted. Default is
'yyyy-mm-dd'.}

\item{integer_format}{How the integer columns should be formatted. Default is
'#,###0'.}

\item{double_format}{How the double columns should be formatted. Default is
'#,###.00'.}
}
\value{
A nested data.table indicating the name of the table and the column
  formats.
}
\description{
This function defines the Power BI column properties of a data frame.
}
