% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{pbi_schema_create}
\alias{pbi_schema_create}
\title{Power BI dataset schema}
\usage{
pbi_schema_create(
  dt_list,
  dataset_name = "My Power BI Dataset",
  table_name_list,
  date_format = "yyyy-mm-dd",
  integer_format = "#,###0",
  double_format = "#,###.00",
  default_mode = c("Push", "Streaming", "PushStreaming", "AsOnPrem", "AsAzure")
)
}
\arguments{
\item{dt_list}{A list of data frames}

\item{dataset_name}{The name of the Power BI dataset.}

\item{table_name_list}{A list of names corresponding to the list of data
frames.}

\item{date_format}{The format of the date columns (if any). Default is
'yyyy-mm-dd'.}

\item{integer_format}{The format of the integer columns (if any). Default is
'#,###0'.}

\item{double_format}{The format of the double columns (if any). Default is
'#,###.00'.}

\item{default_mode}{The dataset mode or type.}
}
\value{
A list with schema properties.
}
\description{
Creates a Power BI push dataset schema
}
\examples{
pbi_schema_create(
  dt_list = list(iris, mtcars),
  dataset_name = "An iris and mtcars dataset",
  table_name_list = list("iris", "mtcars")
  )
}
