% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{pbi_schema_relation_create}
\alias{pbi_schema_relation_create}
\title{Title}
\usage{
pbi_schema_relation_create(
  from_table = NULL,
  from_column = NULL,
  to_table = NULL,
  to_column = from_column,
  direction = c("OneDirection", "BothDirections", "Automatic"),
  name = paste0(from_table, to_table, from_column)
)
}
\arguments{
\item{from_table}{The from table}

\item{from_column}{The from column}

\item{to_table}{The to table}

\item{to_column}{The to columns}

\item{direction}{The crossfiltering direction}

\item{name}{The name}
}
\value{
A data.table
}
\description{
My description
}
\examples{
# An example
}
