% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{pbi_schema_table_prop}
\alias{pbi_schema_table_prop}
\title{Set Power BI table properties}
\usage{
pbi_schema_table_prop(
  dt,
  date_format = "yyyy-mm-dd",
  integer_format = "#,###0",
  double_format = "#,###.00"
)
}
\arguments{
\item{dt}{A data frame.}

\item{date_format}{The Power BI format of date columns. Default is 'yyyy-mm-dd'.}

\item{integer_format}{The Power BI format of integer columns. Default is '#,###0'.}

\item{double_format}{The Power BI format of double columns. Default is '#,###.00'.}
}
\value{
A nested data.table.
}
\description{
This function defines the Power BI table properties of a data frame.
}
\examples{
pbi_schema_table_prop(iris)
}
