% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbi_functions.R
\name{pbi_schema_types_infer}
\alias{pbi_schema_types_infer}
\title{Infer the data type of a vector}
\usage{
pbi_schema_types_infer(dt)
}
\arguments{
\item{dt}{A vector of type POSIXt, Factor, Logical, Double or Integer}
}
\value{
A string indicating the type of the vector
}
\description{
Infer the data type of a vector
}
\examples{
sapply(iris, pbi_schema_types_infer)
}
