% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dim_channel}
\alias{dim_channel}
\title{Demo data: Dim Channel}
\format{
A data frame with 616 rows and 5 columns:
\describe{
  \item{channel_key}{customer ID, factor vector}
  \item{channel_grouping}{transaction date1}
  \item{source}{transaction date2}
  \item{medium}{transaction date3}
  \item{campaign}{transaction date4}
  }
}
\source{
Google Analytics
}
\usage{
dim_channel
}
\description{
Anonymized data from google analytics.
}
\keyword{datasets}
