% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dim_market}
\alias{dim_market}
\title{Demo data: Dim Market}
\format{
A data frame with 2 rows and 3 columns:
\describe{
  \item{market_key}{customer ID, factor vector a}
  \item{view_name}{transaction date1 b}
  \item{country}{transaction date2 c}
  }
}
\source{
Google Analytics
}
\usage{
dim_market
}
\description{
Anonymized data from google analytics.
}
\keyword{datasets}
