% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fact_websales}
\alias{fact_websales}
\title{Fact Web Sales.}
\format{
A data frame with 10483 rows and 2 variables:
\describe{
  \item{date}{customer ID, factor vector a}
  \item{market_key}{transaction date2 c}
  \item{channel_key}{transaction date2 c}
  \item{impressions}{transaction date1 b}
  \item{ad_clicks}{transaction date2 c}
  \item{ad_cost}{customer ID, factor vector a}
  \item{sessions}{transaction date1 b}
  \item{transactions}{transaction date2 c}
  \item{revenue}{transaction date2 c}
  }
}
\source{
Google Analytics
}
\usage{
fact_websales
}
\description{
Anonymized data from google analytics.
}
\keyword{datasets}
