% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/authentication.R
\name{pbi_auth}
\alias{pbi_auth}
\title{Authenticate your Power BI API session}
\usage{
pbi_auth(
  tenant = Sys.getenv("PBI_TENANT"),
  app = Sys.getenv("PBI_APP"),
  password = Sys.getenv("PBI_PW")
)
}
\arguments{
\item{tenant}{Your Microsoft tenant ID}

\item{app}{Your Microsoft App ID}

\item{password}{Your Microsoft App password}
}
\value{
An active token
}
\description{
Supply your application name, client ID and client secret. You will need to
register an application here:
https://powerbi.microsoft.com/en-us/documentation/powerbi-developer-register-a-web-app/
}
