% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{pbi_delete_dataset}
\alias{pbi_delete_dataset}
\title{Delete dataset}
\usage{
pbi_delete_dataset(group_id, dataset_id)
}
\arguments{
\item{group_id}{The dataset ID.}

\item{dataset_id}{The workspace ID.}
}
\value{
Deletes the entire dataset.
}
\description{
Deletes the specified dataset from the specified workspace. Applicable to
push datasets as well as imported datasets.
}
\examples{

\dontrun{

group_id <- "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
dataset_id <- "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"

pbi_delete_dataset(group_id, dataset_id)
}
}
