% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete.R
\name{pbi_delete_rows}
\alias{pbi_delete_rows}
\title{Delete rows}
\usage{
pbi_delete_rows(group_id, dataset_id, table_name)
}
\arguments{
\item{group_id}{The Power BI workspace ID.}

\item{dataset_id}{The Power BI dataset ID.}

\item{table_name}{The Power BI table name.}
}
\value{
All rows will be deleted from the specified table.
}
\description{
Deletes all rows from the specified table within the specified dataset
from the specified workspace (group ID). Only applicable to push datasets.
}
\examples{

\dontrun{

group_id <- "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
dataset_id <- "xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx"
table_name <- "My Table"

pbi_delete_rows(group_id, dataset_id, table_name)
}
}
