% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list.R
\name{pbi_list_groups}
\alias{pbi_list_groups}
\title{Get a list of workspaces}
\usage{
pbi_list_groups()
}
\value{
A data frame with workspaces.
}
\description{
Returns the ids and meta data of all Power BI workspaces to which the service
principal app has been granted access.
}
\examples{

\dontrun{

pbi_list_groups()
}
}
