% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/schema.R
\name{pbi_schema_create}
\alias{pbi_schema_create}
\title{Create a Power BI dataset schema}
\usage{
pbi_schema_create(
  dt_list,
  dataset_name = "My Power BI Dataset",
  table_name_list,
  relations_list = NULL,
  date_format = "yyyy-mm-dd",
  integer_format = "#,###0",
  double_format = "#,###.00",
  sort_by_col = NULL,
  hidden_col = NULL,
  default_mode = c("Push", "Streaming", "PushStreaming", "AsOnPrem", "AsAzure")
)
}
\arguments{
\item{dt_list}{A list of data frames which the schema should be inferred from.}

\item{dataset_name}{A custom name of the Power BI dataset.}

\item{table_name_list}{A list of custom names corresponding to the list of data
frames.}

\item{relations_list}{A list of relation definitions returned by
pbi_schema_relation_create()}

\item{date_format}{The format of date columns (if any). Default is
'yyyy-mm-dd'.}

\item{integer_format}{The format of integer columns (if any). Default is
'#,###0'.}

\item{double_format}{The format of double columns (if any). Default is
'#,###.00'.}

\item{sort_by_col}{List of cols to be sorted.}

\item{hidden_col}{List of cols to be hidden.}

\item{default_mode}{The dataset mode or type. Defaults to 'Push'.}
}
\value{
A list with schema properties.
}
\description{
Creates a Power BI dataset schema from a set of data frames. Columns and data
types will be inferred from the data frames.  Only applicable to push
datasets.
}
\examples{
pbi_schema_create(
  dt_list = list(iris, mtcars),
  dataset_name = "An iris and mtcars dataset",
  table_name_list = list("iris", "mtcars")
  )
}
