% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{pbi_schema_types_infer}
\alias{pbi_schema_types_infer}
\title{Get Power BI data type}
\usage{
pbi_schema_types_infer(vector)
}
\arguments{
\item{vector}{A vector of type POSIXt, Date, Factor, Logical, Double, Integer, Numeric or Character.}
}
\value{
A string indicating the Power BI data type of the vector.
}
\description{
This function detects an R vector's class and converts it into the equivalent data type in Power BI.
}
\examples{
sapply(iris, pbi_schema_types_infer)
}
